RollerDao
1. Project category. 
Governance - DAOs

2. Github Link.
https://github.com/BetX19/RollerDao 

3. Video Link.



4.Team name. 
The Rollers

5. Project description.

Roller DAO is a crowdfunding platform, built and executed as a Roll up in the Cartesi ecosystem. 
Roller DAO allows creators to look for funding to develop their projects. In order to do this, they submit a summary of how the project is going to work and the necessary steps to make it happen.  They set their project’s funding goal and deadline. 
Users in the role of sponsors or backers, are able to support creators' ideas, by linking their wallets and locking their tokens to help make the idea happen, within the deadline previously established. 
If the project succeeds in reaching its funding goal, tokens will be given to the creator in different stages, stated previously and according to the project development. Sponsors will be rewarded with benefits by creators. This will be beforehand established by creators.
If the project doesn’t succeed in reaching its funding goal, tokens will be returned to backers.


Creators will be allowed to upload their projects to the platform, projects will be pre listed and after going through vote by members of the community/Dao, they will be listed in the platform  if the vote is positive or they will go to a waiting list, if negative.

 

6. Names of the participants.
Mayurch - mayurchougule1@gmail.com - https://github.com/mmchougule  
Maximiliano Provenzano - maxi.prudenzano91@gmail.com - 
Matias Guagliardo - matias.guagliardo@gmail.com - https://github.com/Mattteus1 
Gustavo Matta. - linkzelda2016o@gmail.com - https://github.com/BetX19 


7.Project name.
Roller DAO

8. Project logo.




9.Problem description. 

The problem can be summarized in the lack of decentralized spaces to connect creators and sponsors or backers, to work together in developing and launching new products and services to the market. 


10.Solution or how the project helps to solve the problem. 

Roller Dao aims to be a tool that unites Creators and Sponsor in a decentralized way. By using the platform creators will be able to submit their ideas to the public and raise funds to achieve their goals. Sponsors will be able to participate in the projects they found interesting, and help them to become a reality. This is possible by linking their virtual wallets, and they can start to interact with the platform, as creators or sponsors.


11. Tech stack


12. Project demo link (page url)